requirements.txt
asttokens==2.2.1
autopep8==2.0.0
azure-core==1.26.3
azure-storage-blob==12.14.1
backcall==0.2.0
bcrypt==4.0.1
blinker==1.5
cachelib==0.9.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
cryptography==39.0.0
decorator==5.1.1
executing==1.2.0
Flask==2.2.2
Flask-Bcrypt==1.0.1
Flask-Caching==2.0.2
Flask-DebugToolbar==0.13.1
Flask-SQLAlchemy==3.0.2
Flask-Uploads==0.2.1
Flask-WTF==1.0.1
greenlet==2.0.1
gunicorn==20.1.0
idna==3.4
importlib-metadata==5.1.0
ipython==8.7.0
isodate==0.6.1
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
MarkupSafe==2.1.1
matplotlib-inline==0.1.6
msrest==0.7.1
oauthlib==3.2.2
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.33
psycopg2-binary==2.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pycodestyle==2.10.0
pycparser==2.21
Pygments==2.13.0
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
SQLAlchemy==1.4.44
stack-data==0.6.2
tomli==2.0.1
traitlets==5.6.0
typing-extensions==4.4.0
urllib3==1.26.13
wcwidth==0.2.5
Werkzeug==2.2.2
WTForms==3.0.1
zipp==3.11.0
